requirements.txt
flask==3.0.2
openai==1.59.9
requests==2.31.0
python-dotenv==1.0.1
pydantic==1.10.13 